refactor(web): move Navbar componente to _app
