+++ apps/web/src/utils/cssClasses.ts
@@ -1,0 +1,1 @@
+export const HOVER_LINK = "hover:font-bold";
